Updated i18n code to work on win32
Currently translations of a pygobject user interface loaded
from a Glade file fail on win32. The cause of this is
currently unknown, but is being investigated by me
and some people upstream. Bajusz Tamás has found a
workaround that loads each interface element and sets the
translation again during the __init__ of the window or
dialog. I implemented this workaround, but it still
needs some testing and cleaning.