Created make win32 script
The script creates the correct project structure, compiles
all the translations and copies them to the correct folder.
Then the script calls Pynsist to compile the win32 installer.
Currently only the 64bit version is built. Future commits
will try to built the installers using the Pynsist-API and
built the 32 and 64 bit verions in one run.