Call function to update the contour range label on init
The layer properties dialog had a bug in non-English locale that
would cause the label to be set as the header of the mo-file.
This only happend under win32 and was probably caused by the
re-translate function not finding a translation for an empty
string. The easiest solution is to just call the method of
the dialog that sets the label according to the state of the
buttons.